+++ lib/addon/hl_alm.tcl
@@ -20,13 +20,13 @@ proc hl_alm::init {w font szfont args} {
   lassign $args clrCOM clrCOMTK clrSTR clrVAR clrCMN clrPROC clrOPT
   dict set font -size $szfont
   dict set font -weight bold;
-  $w tag config tagKEY -font $font -foreground $clrCOM
+  $w tag config almKEY -font $font -foreground $clrCOM
   dict set font -weight normal
-  $w tag config tagPATH -font $font -foreground $clrVAR
-  $w tag config tagARG -font $font -foreground $clrOPT
+  $w tag config almPATH -font $font -foreground $clrVAR
+  $w tag config almARG -font $font -foreground $clrOPT
   dict set font -slant italic
-  $w tag config tagCMNT -font $font -foreground $clrCMN
-  foreach t {KEY PATH ARG CMNT} {after idle $w tag raise tag$t}
+  $w tag config almCMNT -font $font -foreground $clrCMN
+  foreach t {KEY PATH ARG CMNT} {after idle $w tag raise alm$t}
   return [namespace current]::line
 }
 #_______________________
@@ -41,28 +41,28 @@ proc hl_alm::line {w {pos ""} {prevQtd 0}} {
   set il [expr {int($pos)}]
   set line [$w get $il.0 $il.end]
   if {[string trim $line] eq {}} {return yes}
-  foreach t {KEY PATH ARG CMNT} {$w tag remove tag$t $il.0 $il.end}
+  foreach t {KEY PATH ARG CMNT} {$w tag remove alm$t $il.0 $il.end}
   if {[regexp "^\s*#" $line]} {
-    $w tag add tagCMNT $il.0 $il.end
+    $w tag add almCMNT $il.0 $il.end
     return yes
   }
   lassign [regexp -inline {^\s*[^\s]+} $line] lre
   if {$lre ne {}} {
     set p1 [string length $lre]
-    $w tag add tagKEY $il.0 $il.$p1
+    $w tag add almKEY $il.0 $il.$p1
   }
   set path [regexp -inline -all -indices {\s{1}[.]{1}[^\s]+} $line]
   foreach l2 $path {
     lassign $l2 p1 p2
     if {$p1<$p2} {
-      $w tag add tagPATH $il.$p1 $il.[incr p2]
+      $w tag add almPATH $il.$p1 $il.[incr p2]
     }
   }
   set key [regexp -inline -all -indices {\s%.=} $line]
   foreach l2 $key {
     lassign $l2 p1 p2
     if {$p1<$p2} {
-      $w tag add tagARG $il.$p1 $il.[incr p2]
+      $w tag add almARG $il.$p1 $il.[incr p2]
     }
   }
   return yes